Change drawing ink from black to rainbow
This feature fufills the extra credit #1. To ensure this feature is
compatible with extra credit #2 (gradual darkening of cells), the
"filled" class is now used as a flag to assign a random color to a cell
*only* on its first mouse-over; and darkenCell() is adapted to retain
the assigned random color.

The "filled" class was removed from the stylesheet as a result of all
its attributes becoming dynamically assigned in the extra credit, thus
rendering it an empty selector and not needing to be defined.

A helper method to extract rgba values was implemented to keep the code
cleaner. It was not used previously since we were only extracting the
alpha value, whereas now we extract all of r, g, b, alpha.