Improve input validation when generating new grid
Previously, the prompt would ignore any requests to cancel the input:
clicking the Cancel button, pressing the escape key, or refreshing the
page. Validation was also added to reject non-positive numbers (0 and
negatives).

The function to prompt for user input was removed and implemented
directly in the callback function of the button as this more easily
precludes generateGrid() from being called when user input is invalid.